Implement menu-button-customization feature.
This feature allows the user to change the menu button appearance including:
 * the menu button text
 * the menu button icon
 * the size of the icon
 * enable or disable the arrow icon beside the menu button text

Sorry that this commit is so big, but to implement the menu-button-customization feature,
I had to refactor the code base. I hope these changes will help us in the future
to maintain a clean code base.

This commit introduces the following changes:
 * Move the logic for handling settings changes in controller.js, namley in the MenuSettingsController class
 * Move common constants in constants.js
 * Refactor helper.js
 * Refactor some parts in menu.js
 * Implement the basic GUI option for changing the menu button text and the menu button icon
 * Update the schema file to include the new settings options
 * Add the file am.js, which provides customized GTK GUI elements for the prefs.js file

Signed-off-by: Alexander Rüedlinger <a.rueedlinger@gmail.com>

commit