GNOME-Shell 3.25/26: Fix JS warnings
I tested Arc Menu on Ubuntu 17.10 and discovered the following JS warnings:
 * Warning 1: Too many arguments to method Clutter.Actor.add_child: expected 1, got 2
 * Warning 2: That property was defined with 'let' or 'const' inside the module.

It seems that Clutter.Actor.add_child() only expects a single argument, namely a child actor [2].

The fix for warning 1 is pretty simple, we just remove the extra argument.
And to fix warning 2, we simply use the keyword 'var' to export classes and constants to other modules.

Thus, some 'const' keywords are replaced with a 'var' keyword in the following files:
 * controller.js
 * helper.js
 * menu.js
 * constants.js
 * logger.js

[1] http://paste.debian.net/983968
[2] https://developer.gnome.org/clutter/stable/ClutterActor.html#clutter-actor-add-child

Signed-off-by: Alexander Rüedlinger <a.rueedlinger@gmail.com>